Move cache utils to utils :p
